+++ src/test/java/com/mongodb/kafka/connect/sink/processor/field/renaming/RenamerTest.java
@@ -103,4 +103,34 @@ class RenamerTest {
         () -> assertEquals(expectedKeyDoc, sd.getKeyDoc().orElse(new BsonDocument())),
         () -> assertEquals(expectedValueDoc, sd.getValueDoc().orElse(new BsonDocument())));
   }
+
+  @Test
+  @DisplayName("simple field renamer test with auto regexp escaping")
+  void testRenamerUsingRegExpSettingsWithAutoEscaping() {
+    BsonDocument expectedKeyDoc =
+        BsonDocument.parse(
+            "{'FA': 'my field value', 'f2': true, "
+                + "'subDoc': {'FX': 42}, '_F1': {'_F2': 'testing rocks!'}}");
+
+    BsonDocument expectedValueDoc =
+        BsonDocument.parse(
+            "{'abc': 'my field value', 'f2': false, "
+                + "'subDoc': {'123': 0.0}, 'foo_foo_foo': {'_blah__blah_': 23}}");
+
+    MongoSinkTopicConfig cfg =
+        createTopicConfig(
+            FIELD_RENAMER_REGEXP_CONFIG,
+            "[{'regexp': '^key\\..*my.*$', 'pattern': 'my', 'replace': ''},"
+                + "{'regexp': '^key\\..*field.$', 'pattern': 'field', 'replace': 'F'},"
+                + "{'regexp': '^value\\..*$', 'pattern': '\\.', 'replace': '_'}]");
+
+    SinkDocument sd = new SinkDocument(keyDoc, valueDoc);
+    Renamer renamer = new RenameByRegExp(cfg);
+    renamer.process(sd, null);
+
+    assertAll(
+        "key and value doc checks",
+        () -> assertEquals(expectedKeyDoc, sd.getKeyDoc().orElse(new BsonDocument())),
+        () -> assertEquals(expectedValueDoc, sd.getValueDoc().orElse(new BsonDocument())));
+  }
 }