+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - [KAFKA-131](https://jira.mongodb.org/browse/KAFKA-131) Added `copy.existing.pipeline` configuration.
     Note: Allows indexes to be used during the copying process, use when there is any filtering done by the main pipeline.
   - [KAFKA-146](https://jira.mongodb.org/browse/KAFKA-146) Improve error handling and messaging for list configuration options.
+  - [KAFKA-154](https://jira.mongodb.org/browse/KAFKA-154) Improve the handling and error messaging for Json array config values.
 
 
 ## 1.2.0